# dto/admin.go
package dto

import (
	"github.com/e421083458/yqb_monitor/public"
	"github.com/gin-gonic/gin"
	"time"
)

type AdminInfoOutput struct {
	ID           int       `json:"id"`
	Name        string    `json:"name"`
	LoginTime    time.Time `json:"login_time"`
	Avatar       string    `json:"avatar"`
	Introduction string    `json:"introduction"`
	Roles        []string  `json:"roles"`
}

type ChangePwdInput struct {
	Password string `json:"password" form:"password" comment:"密码" example:"123456" validate:"required"` //密码
}

func(params *ChangePwdInput) BindValidparam(c *gin.Context) error {
	return public.DefaultGetValidParams(c, params)
}